Revert "Add bqetl view validate command"
This reverts commit 62c85cb36caadab81f8f426cf81937de8649a80b.
