+++ bigquery_etl/dependency.py
@@ -9,7 +9,7 @@ from typing import Dict, Iterator, List, Tuple
 import click
 import yaml
 
-from bigquery_etl.view.generate_stable_views import get_stable_table_schemas
+from .view.generate_stable_views import get_stable_table_schemas
 
 stable_views = None
